Regroup Gateway feature into Gateway namespace, prepare to Oney add
