+++ src/Resources/config/services.xml
@@ -2,8 +2,7 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <imports>
-        <import resource="services/gateway_factory.xml" />
-        <import resource="services/form.xml" />
+        <import resource="services/gateway.xml" />
         <import resource="services/action.xml" />
     </imports>
