--- src/Resources/config/services/form.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <services>
-        <service id="payplug_sylius_payplug_plugin.form.type.payplug_gateway_configuration" class="PayPlug\SyliusPayPlugPlugin\Form\Type\PayPlugGatewayConfigurationType">
-            <tag name="sylius.gateway_configuration_type" type="payplug" label="payplug_sylius_payplug_plugin.ui.payplug_gateway_label" />
-            <tag name="form.type" />
-            <argument type="service" id="translator"/>
-        </service>
-    </services>
-</container>
