--- src/Resources/config/services/gateway_factory.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <services>
-        <service id="payplug_sylius_payplug_plugin.gateway_factory.payplug" class="Payum\Core\Bridge\Symfony\Builder\GatewayFactoryBuilder">
-            <argument type="string">PayPlug\SyliusPayPlugPlugin\PayPlugGatewayFactory</argument>
-            <tag name="payum.gateway_factory_builder" factory="payplug" />
-        </service>
-    </services>
-</container>
